+++ search-films/src/components/Flex/index.jsx
@@ -1,5 +1,5 @@
 import { default as CssUtils } from '../shared/CssUtils';
-import './styles/index.css';
+import styles from './styles/index.module.css';
 /**
  * util Flex component 
  * @param {string} position - (byDefault=vertical) vertical/horizontal flex container
@@ -14,10 +14,10 @@ function Flex(props) {
 	const paddingTop = props.paddingTop ? CssUtils.addPxToCssProperty(props.paddingTop) : /*byDefault=*/ 0;
 
 	const flexContainer = position == 'vertical' ?
-		<div style={{paddingTop}} className='flex-container-vertical'>
+		<div style={{paddingTop}} className={styles['flex-container-vertical']}>
 			{props.children}
 		</div> :
-		<div style={{paddingTop}} className='flex-container-horizontal'>
+		<div style={{paddingTop}} className={styles['flex-container-horizontal']}>
 			{props.children}
 		</div>;