+++ search-films/src/components/Paragraph/index.jsx
@@ -1,5 +1,5 @@
 import { default as CssUtils } from '../shared/CssUtils.js';
-import './styles/index.css';
+import styles from './styles/index.module.css';
 
 function Paragraph({text, fontSizeInPx}) {
 	// здесь стили вот так задаются, чтобы на маленьких экранах было норм
@@ -7,7 +7,7 @@ function Paragraph({text, fontSizeInPx}) {
 		fontSize: CssUtils.addPxToCssProperty(fontSizeInPx)
 	};
 	return (
-		<p className="paragraph" style={style}>{text}</p>
+		<p className={styles.paragraph} style={style}>{text}</p>
 	);
 }